+++ HISTORY.md
@@ -1,7 +1,13 @@
 # History.md
 
+## v0.12.0 2025-02-04
+- Add parameter `textcolor` to:
+    - `angulardimension`
+    - `phasordimension`
+    - `lengthdimension`
+
 ## v0.11.0 2024-12-13
-- Add parameter gapcolor to `phasorsine_hline` and `phasorcosine_hline`
+- Add parameter `gapcolor` to `phasorsine_hline` and `phasorcosine_hline`
 
 ## v0.10.4 2024-09-14
 - Remove obsoletely plotted lines